infra/requirements-cog.txt
# COG (Cloud Optimized GeoTIFF) Dependencies
# Install with: pip install -r infra/requirements-cog.txt

# Core geospatial libraries
rasterio>=1.3.0
rio-cogeo>=3.0.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0

# Planetary image support (optional)
planetaryimage>=0.5.0

# Additional GDAL drivers (optional)
# Uncomment if you need specific drivers
# gdal[all]>=3.4.0
